Split up the main.cpp file into respective subclass (headers for now)
