+++ BCMHost.hh
@@ -1,0 +1,10 @@
+class BCMHost
+{
+public:
+	BCMHost() {
+		bcm_host_init();
+	};
+	~BCMHost() {
+		bcm_host_deinit();
+	}
+};
